Refresh kitchen tests
Former tests were not doing much.
- Setup dokken tests for workstations + local one for the CI.
- Add inspec tests to check that the attributes are computed

It uses local tests in the CI to "simplify" the tests on Windows by
simply running kitchen on the GA runner.
In the end on Windows it is never simple :D some trick has to be done
to ensure the path is not too long and the env is clean.

Notes:
- Attributes computation depends on the platform, so only the "format"
  is verified